requirements.txt
clip
colour
easydict
einops
EVM
faiss-gpu
h5py
huggingface_hub
matplotlib
MinkowskiEngine
numpy
omegaconf
open3d
open3d_python
open_clip_torch
pandas
plyfile
scikit_learn
scipy
setuptools
timm
torch==2.1.2+cu118
torch_redstone==0.0.6
torchvision==0.16.2+cu118
tqdm
trimesh
